Cleaned up the code

+++ termget.py
@@ -9,40 +9,26 @@ package_file_read = open("/home/" + getpass.getuser() + "/.termget/termget-packa
 credit = "TermGet was created by:\n- PizzaLovingNerd (main developer)\n- SudoLinux"
 
 if len(sys.argv) == 2:
-    if sys.argv[1] == "apt-get":
-        package = "apt-get"
-    elif sys.argv[1] == "pacman":
-        package = "pacman"
-    elif sys.argv[1] == "xbps":
-        package = "xbps"
-    elif sys.argv[1] == "dnf":
-        package = "dnf"
-    elif sys.argv[1] == "zypper":
-        package = "zypper"
-    elif sys.argv[1] == "eopkg":
-        package = "eopkg"
-    elif sys.argv[1] == "pip":
-        package = "pip"
+    if sys.argv[1] == "apt-get": package = "apt-get"
+    elif sys.argv[1] == "pacman": package = "pacman"
+    elif sys.argv[1] == "xbps": package = "xbps"
+    elif sys.argv[1] == "dnf": package = "dnf"
+    elif sys.argv[1] == "zypper": package = "zypper"
+    elif sys.argv[1] == "eopkg": package = "eopkg"
+    elif sys.argv[1] == "pip": package = "pip"
 
 if package == " ":
-    if package_file_read == "apt-get":
-        package = "apt-get"
-    elif package_file_read == "pacman":
-        package = "pacman"
-    elif package_file_read == "xbps":
-        package = "xbps"
-    elif package_file_read == "dnf":
-        package = "dnf"
-    elif package_file_read == "zypper":
-        package = "zypper"
-    elif package_file_read == "eopkg":
-        package = "eopkg"
+    if package_file_read == "apt-get": package = "apt-get"
+    elif package_file_read == "pacman": package = "pacman"
+    elif package_file_read == "xbps": package = "xbps"
+    elif package_file_read == "dnf": package = "dnf"
+    elif package_file_read == "zypper": package = "zypper"
+    elif package_file_read == "eopkg": package = "eopkg"
     print("package manager set to " + package)
 
 #Checks for command line argument
 
-def clear():
-    os.system("clear")
+def clear(): os.system("clear")
     #Runs 'clear' over shell to clear the screen.
 
 clear()
@@ -105,22 +91,14 @@ if package != "pip":
             clear()
             user = input("Please enter search query: ")
             print(" ")
-            if package == "apt-get":
-                os.system("sudo apt-cache search " + user)
-            elif package == "pacman":
-                user1 = input("Which package manager would you like to use?\n\n1. pacman\n2. yaourt\n")
-            if user1 == "1":
-                os.system("sudo pacman -Ss " + user)
-            if user1 == "2":
-                os.system("yaourt -Ss " + user)
-            elif package == "xbps":
-                os.system("sudo xbps-query -Rs " + user)
-            elif package == "dnf":
-                os.system("sudo dnf search " + user)
-            elif package == "zypper":
-                os.system("sudo zypper search " + user)
-            elif package == "eopkg":
-                os.system("eopkg search " + user)
+            if package == "apt-get": os.system("sudo apt-cache search " + user)
+            elif package == "pacman": user1 = input("Which package manager would you like to use?\n\n1. pacman\n2. yaourt\n")
+            if user1 == "1": os.system("sudo pacman -Ss " + user)
+            if user1 == "2": os.system("yaourt -Ss " + user)
+            elif package == "xbps": os.system("sudo xbps-query -Rs " + user)
+            elif package == "dnf": os.system("sudo dnf search " + user)
+            elif package == "zypper": os.system("sudo zypper search " + user)
+            elif package == "eopkg": os.system("eopkg search " + user)
             input("\nPress enter to continue")
 
         if user == "2": #Install
@@ -128,22 +106,15 @@ if package != "pip":
             user = input("Please enter which package(s) to install: ")
             print("")
 
-            if package == "apt-get":
-                os.system("sudo apt-get install " + user)
+            if package == "apt-get": os.system("sudo apt-get install " + user)
             if package == "pacman":
                 user1 = input("Which package manager would you like to use?\n\n1. pacman\n2. yaourt\n")
-                if user1 == "1":
-                    os.system("sudo pacman -S " + user)
-                if user1 == "2":
-                    os.system("yaourt -S " + user)
-            if package == "xbps":
-                os.system("sudo xbps-install " + user)
-            if package == "dnf":
-                os.system("sudo dnf install " + user)
-            if package == "zypper":
-                os.system("sudo zypper install " + user)
-            if package == "eopkg":
-                os.system("sudo eopkg install " + user)
+                if user1 == "1": os.system("sudo pacman -S " + user)
+                if user1 == "2": os.system("yaourt -S " + user)
+            if package == "xbps": os.system("sudo xbps-install " + user)
+            if package == "dnf": os.system("sudo dnf install " + user)
+            if package == "zypper": os.system("sudo zypper install " + user)
+            if package == "eopkg": os.system("sudo eopkg install " + user)
             input("\nPress enter to continue")
 
         if user == "3": #Remove
@@ -153,20 +124,13 @@ if package != "pip":
             if package == "apt-get":
                 user1 = input("How will you like to remove the package?\n\n1. remove, removes just the package (faster)\n2. purge, removes the package, and all it's configuration files (saves space)")
                 clear()
-                if user1 == "1":
-                    os.system("sudo apt-get remove " + user)
-                if user1 == "2":
-                    os.system("sudo apt-get purge " + user)
-            if package == "pacman":
-                os.system("sudo pacman -Rs " + user)
-            if package == "xbps":
-                os.system("sudo xbps-remove " + user)
-            if package == "dnf":
-                os.system("sudo dnf remove " + user)
-            if package == "zypper":
-                os.system("sudo zypper remove " + user)
-            if package == "eopkg":
-                os.system("sudo eopkg remove " + user)
+                if user1 == "1": os.system("sudo apt-get remove " + user)
+                if user1 == "2": os.system("sudo apt-get purge " + user)
+            if package == "pacman": os.system("sudo pacman -Rs " + user)
+            if package == "xbps": os.system("sudo xbps-remove " + user)
+            if package == "dnf": os.system("sudo dnf remove " + user)
+            if package == "zypper": os.system("sudo zypper remove " + user)
+            if package == "eopkg": os.system("sudo eopkg remove " + user)
             input("\nPress enter to continue")
 
         if user == "4": #Updates Packages
@@ -175,35 +139,26 @@ if package != "pip":
             if package == "apt-get":
                 os.system("sudo apt-get upgrade")
                 os.system("sudo apt-get dist-upgrade")
-            if package == "pacman":
-                os.system("sudo pacman -Syu")
-            if package == "xbps":
-                os.system("sudo xbps-install -Su")
+            if package == "pacman": os.system("sudo pacman -Syu")
+            if package == "xbps": os.system("sudo xbps-install -Su")
             if package == "dnf":
                 os.system("sudo dnf upgrade")
                 os.system("sudo dnf distro-sync")
             if package == "zypper":
                 os.system("sudo zypper update zypper up")
                 os.system("sudo zypper dup")
-            if package == "eopkg":
-                os.system("sudo eopkg upgrade")
+            if package == "eopkg": os.system("sudo eopkg upgrade")
             input("\nPress enter to continue")
 
         if user == "5": #Updates Repo
             clear()
             print("\n")
-            if package == "apt-get":
-                os.system("sudo apt-get update")
-            elif package == "pacman":
-                os.system("sudo pacman -Sy")
-            elif package == "xbps":
-                os.system("sudo xbps-install -S")
-            elif package == "dnf":
-                os.system("sudo dnf clean expire-cache && dnf check-update")
-            elif package == "zypper":
-                os.system("sudo zypper refresh zypper ref")
-            elif package == "eopkg":
-                os.system("sudo eopkg ur")
+            if package == "apt-get": os.system("sudo apt-get update")
+            elif package == "pacman": os.system("sudo pacman -Sy")
+            elif package == "xbps": os.system("sudo xbps-install -S")
+            elif package == "dnf": os.system("sudo dnf clean expire-cache && sudo dnf check-update")
+            elif package == "zypper": os.system("sudo zypper refresh zypper ref")
+            elif package == "eopkg": os.system("sudo eopkg ur")
             input("\nPress enter to continue")
 
         if user == "6": #Cleans
@@ -235,8 +190,7 @@ if package != "pip":
             time.sleep(3)
             #If you contribute, please add your name.
 
-        if user == "8": #Quit
-            quit()
+        if user == "8": quit()
 
             
 if package == "pip": #Starts a loop
@@ -278,5 +232,4 @@ if package == "pip": #Starts a loop
 		    print(credit)
 		    time.sleep(3)
 		
-	    if user == "6": #Quit
-		    quit()
+	    if user == "6": quit()
